fix: 예약생성 동시성테스트를 인수테스트에서 통합테스트로 수정
